Move Generator tests classes to appropriate package
